Principal components of RegMap data on A. thaliana

Data on 1,307 Arabidopsis thaliana accessions collected around the world. Data include geographic co-ordinates and results of applying principal components analysis (PCA) to genotypes at 214,051 single nucleotide polymorphisms (SNPs). This data is obtained from Novembre, et al.2008.See PCAviz package.

data("regmappcs")

Format

regmappcs is a data frame with 1,307 rows and 21 columns, in which each row represents an A. thaliana accession included in the RegMap panel. The last 11 columns (labeled "PC1" through "PC10") store the projection of the samples onto the first 10 PCs.

Source

Arabidopsis thaliana SNP genotypes were obtained from the website of Joy Bergelson's lab at the University of Chicago, http://bergelson.uchicago.edu/?page_id=790. We used the data from Novembre, et al.2018 PCAviz package.

Details

PCs were computed using rpca from the rsvd package after first converting the genotypes to a 1,307 x 214,051 binary matrix.

References

Horton, M. W., Hancock, A. M., Huang, Y. S., Toomajian, C., Atwell, S., Auton, A., et al (2012). Genome-wide patterns of genetic variation in worldwide Arabidopsis thaliana accessions from the RegMap panel. Nature Genetics 44, 212–216. Novembre, J., R. Williams, H. Pourreza, Y. Wang, and P. Carbonetto.(2018) "PCAviz: Visualizing principal components analys." R Package.

Examples


# View the first few data samples.
data(regmappcs)
head(regmappcs)

#>   array_id ecotype_id median_intensity latitude longitude nativename firstname
#> 1      641       9381              476  55.8106   14.2091      Fri 1    Alison
#> 2      642       9380              556  55.7488   13.3742     FlyA 3    Alison
#> 3      643       9378              542  63.0165   18.3174      FaU 4    Alison
#> 4      644       9371              478  63.0160   18.3175      FaL 1    Alison
#> 5      645       9367              560  57.6787   14.9981      EkN 3    Alison
#> 6      646       9363              363  62.9147   18.4045      EdJ 2    Alison
#>     surname            site   region country      PC1      PC2     PC3     PC4
#> 1 Anastasio       Frisaboda S Sweden     SWE -27.7564  34.4838  6.2580 -2.0400
#> 2 Anastasio         Flyinge S Sweden     SWE -20.1760  34.9538 15.6064 -6.1097
#> 3 Anastasio  Faberget-Upper N Sweden     SWE -56.7601  89.7009  9.0919 11.4833
#> 4 Anastasio  Faberget-Lower N Sweden     SWE -55.6441  88.5797  8.0689  9.9808
#> 5 Anastasio Eksjoholm-Norra C Sweden     SWE -38.5506   6.1973 -6.9958 -5.7941
#> 6 Anastasio    Edsattersjon N Sweden     SWE -60.6846 109.8771 16.1302 13.5530
#>        PC5      PC6      PC7      PC8      PC9     PC10
#> 1  -6.5748  -1.4439   2.2803  25.6212 -10.9108  18.6137
#> 2 -99.2813 -44.4414 -38.4179 -40.0077  24.6401   6.5482
#> 3  46.6343  -1.4997   0.0836 -14.2385  -3.0902  -6.1308
#> 4  46.1855  -1.2229   1.5694 -15.3463  -2.3640  -4.1847
#> 5 -12.8444   1.4173   5.9657  23.0560  -3.2038 -23.0929
#> 6  46.3648  -2.1523 -10.4390 -20.7712  10.2359  -1.0296


if (FALSE) {
# See the vignette for more examples.
vignette("regmappcs")
}